Move error revalidator to separate file.
